move TraceSwitchExtensions to Primitives
