+++ src/System.Windows.Forms.Primitives/src/System/Diagnostics/TraceSwitchExtensions.cs
@@ -5,7 +5,7 @@ using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Text;
 
-namespace System.Windows.Forms;
+namespace System.Diagnostics;
 
 internal static class TraceSwitchExtensions
 {
