+++ src/System.Windows.Forms/src/System.Windows.Forms.csproj
@@ -43,7 +43,6 @@
     <Compile Include="..\..\Common\src\IComparerHelpers.cs" Link="Common\IComparerHelpers.cs" />
     <Compile Include="..\..\Common\src\Obsoletions.cs" Link="Common\Obsoletions.cs" />
     <Compile Include="..\..\Common\src\RTLAwareMessageBox.cs" Link="Common\RTLAwareMessageBox.cs" />
-    <Compile Include="..\..\Common\src\TraceSwitchExtensions.cs" Link="Common\TraceSwitchExtensions.cs" />
   </ItemGroup>
 
   <ItemGroup>
