+++ src/System.Windows.Forms/src/System/Resources/ResXResourceWriter.cs
@@ -3,7 +3,6 @@
 
 using System.Reflection;
 using System.Text;
-using System.Windows.Forms;
 using System.Xml;
 
 namespace System.Resources;
